Crate telamon_gen

Modules

ast
Syntaxic tree for the constraint description.

error

ir
The constraint description for the ir.

lexer

parser

Functions

process
Parses a constraint description file.

process_file
Process a file and stores the result in an other file. This is meant to be used from build.rs files and may print output to be interpreted by cargo on stdout.