Module telamon_gen::ast

Syntaxic tree for the constraint description.

Structs

Ast

ChoiceInstance
A reference to a choice instantiated with the given variables.

Constraint
A constraint that must be enforced by the IR.

CounterBody

CounterDef

EnumDef
A toplevel definition or constraint.

IntegerDef
A toplevel integer

Position

Quotient

SetDef

SetRef
References a set.

Spanned

VarDef
Defines a variable.

Enums

ChoiceDef

Condition
One of the condition that has to be respected by a constraint.

CounterVal
The value of a counter increment.

EnumStatement
A statement in an enum definition.

Hint
Hint is a token representation.

Statement
A toplevel definition or constraint.

Symmetry
Indicates if an enum exhibits symmetry.

TypeError
TypeEror is the error representation of telamon's.