Enum telamon_gen::ir::ChoiceAction

pub enum ChoiceAction {
    FilterSelf,
    Filter {
        choice: ChoiceInstance,
        filter: FilterCall,
    },
    IncrCounter {
        counter: ChoiceInstance,
        value: CounterVal,
        incr_condition: ValueSet,
    },
    UpdateCounter {
        counter: ChoiceInstance,
        incr: ChoiceInstance,
        incr_condition: ValueSet,
    },
    Trigger {
        id: usize,
        condition: ChoiceCondition,
        code: Code,
        inverse_self_cond: bool,
    },
}

An action to perform,

Variants
FilterSelf

The choice runs all its filters on itself.

Filter

The choice runs a filter on another choice.

Fields of Filter

choice: ChoiceInstance

filter: FilterCall

IncrCounter

Increments a counter if the increment condition is statisfied.

Fields of IncrCounter

counter: ChoiceInstance

value: CounterVal

incr_condition: ValueSet

UpdateCounter

Update a counter after the increment value is changed.

Fields of UpdateCounter

counter: ChoiceInstance

incr: ChoiceInstance

incr_condition: ValueSet

Trigger

Triggers a lowering.

Fields of Trigger

id: usize

condition: ChoiceCondition

code: Code

inverse_self_cond: bool

Methods

impl ChoiceAction

Important traits for Box<W>
impl<W> Write for Box<W> where
W: Write + ?Sized, impl<R> Read for Box<R> where
R: Read + ?Sized, impl<I> Iterator for Box<I> where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn variables<'a>(&'a self) -> Box<Iterator<Item = &'a Set> + 'a>

Returns the list of variables to allocate.

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn inputs(&self) -> &[ChoiceInstance]

Returns the list of inputs used by the action.

pub fn inverse_self(&mut self, ir_desc: &IrDesc)

Inverse references to the value of the choice the action is registered in.