Enum telamon_gen::ir::ChoiceArguments

pub enum ChoiceArguments {
    Plain {
        vars: Vec<(RcStr, Set)>,
    },
    Symmetric {
        names: [RcStr; 2],
        t: Set,
        inverse: bool,
    },
}

Defines the parameters for which the Choice is defined.

Variants
Plain

The Choice is defined for all comibnation of variables of the given sets Each variable can only appear once.

Fields of Plain

vars: Vec<(RcStr, Set)>

Symmetric

The Choice is defined on a triangular space. The rests is obtained by symmetry.

Fields of Symmetric

names: [RcStr; 2]

t: Set

inverse: bool

Methods

impl ChoiceArguments

pub fn new(vars: Vec<(RcStr, Set)>, symmetric: bool, inverse: bool) -> Self

Creates a new ChoiceArguments.

pub fn names<'a>(&'a self) -> <'a>

Returns the name of the arguments.

pub fn sets<'a>(&'a self) -> <'a>

Returns the sets of the arguments.

pub fn get(&self, index: usize) -> (&RcStr, &Set)

Returns the name and set of the argument at the given position.

pub fn iter<'a>(&'a self) -> <'a>

Iterates over the arguments, with their sets and names.

pub fn is_symmetric(&self) -> bool

Indicates if the arguments iteration domain is triangular.

pub fn len(&self) -> usize

Returns the number of arguments.