Enum telamon_gen::ir::ValueType

pub enum ValueType {
    Enum(RcStr),
    Range {
        is_half: bool,
    },
    NumericSet(Code),
    Constant,
}

Specifies the type of the values a choice can take.

Variants
Enum(RcStr)

Generated type that represents the values of an enum choice.

Range

Represents a contiguous range of values.

Fields of Range

is_half: bool
NumericSet(Code)

Represents a small set of integers.

Constant

Represents an external constant, provided by the user. Its exact type is unspecified.

Methods

impl ValueType

pub fn full_type(self) -> Self

Returns the full type, instead of a the trimmed one.

pub fn as_enum(&self) -> Option<&RcStr>

Returns the enum name, if applicable.