Struct telamon_gen::ir::Adaptator

pub struct Adaptator { /* fields omitted */ }

Represent a transformation to apply to a rule to fir it in a new context.

Methods

impl Adaptator

pub fn from_arguments(args: &[Variable]) -> Self

Creates an adaptator that maps the arguments to the given names.

pub fn variable(&self, var: Variable) -> Variable

Adapts a variable.

pub fn input(&self, input: usize) -> (usize, bool)

Returns the new ID of an input and indicates if it is inversed.

pub fn set_variable(&mut self, old: Variable, new: Variable) -> Option<Variable>

Sets the mapping of a variable. Returns the previous mapping.

pub fn set_input(&mut self, old: usize, new: usize)

Sets the mapping of an input.

pub fn set_inversed(&mut self, input: usize)

Sets an input as inversed.