Struct telamon_gen::ir::Enum

pub struct Enum { /* fields omitted */ }

A choice that can take a few predefined values.

Methods

impl Enum

pub fn new(
name: RcStr,
doc: Option<RcStr>,
inverse: Option<Vec<(RcStr, RcStr)>>
) -> Self

Creates a new enum definition.

pub fn name(&self) -> &RcStr

Returns the name of the enum.

pub fn add_value(&mut self, name: RcStr, doc: Option<String>)

Adds a possible value to the enum.

pub fn add_alias(
&mut self,
name: RcStr,
vals: HashSet<RcStr>,
doc: Option<String>
)

Adds an alias to the enum possible values.

pub fn aliases(&self) -> &IndexMap<RcStr, (HashSet<RcStr>, Option<String>)>

Lists the aliases.

pub fn doc(&self) -> Option<&str>

Returns the documentation associated with the enum.

pub fn values(&self) -> &IndexMap<RcStr, Option<String>>

Returns the values the enum can take, and their associated comment.

pub fn expand<IT: IntoIterator<Item = RcStr>>(&self, set: IT) -> HashSet<RcStr>

Replaces aliases by the corresponding values.

pub fn inverse<'a>(&'a self, value: &'a RcStr) -> &'a RcStr

Inverse an antisymmetric value.

pub fn inverse_mapping(&self) -> Option<&[(RcStr, RcStr)]>

Returns the mapping to apply to obtain the symmetric of a value.