Struct telamon_gen::ir::IrDesc

pub struct IrDesc { /* fields omitted */ }

Describes the choices that constitute the IR.

Methods

impl IrDesc

pub fn add_choice(&mut self, choice: Choice)

Adds a Choice to the IR description.

pub fn add_enum(&mut self, enum_: Enum)

Adds an Enum to the IR desctiption.

pub fn choices<'a>(&'a self) -> <'a>

List the choice definitions.

pub fn enums<'a>(&'a self) -> <'a>

List the enum definitions.

pub fn get_enum<'a>(&'a self, name: &str) -> &'a Enum

Returns the enum with the given name.

pub fn get_choice<'a>(&'a self, name: &str) -> &'a Choice

Returns the choice with the given name.

pub fn set_defs<'a>(&'a self) -> <'a>

Iterates over all the sets.

pub fn add_set_def(&mut self, def: Rc<SetDef>)

Register a set definition.

pub fn get_set_def<'a>(&'a self, name: &str) -> &'a Rc<SetDef>

Returns the set definition associated with a name.

pub fn add_filter(
&mut self,
choice: RcStr,
filter: Filter,
forall_vars: Vec<Set>,
set_constraints: SetConstraints
)

Adds a filter to a choice.

pub fn add_onchange(&mut self, choice: &str, action: OnChangeAction)

pub fn add_trigger(&mut self, trigger: Trigger) -> usize

Adds a trigger to a choice.

pub fn triggers(&self) ->

Iterates on the triggers.

pub fn adapt_env(
&self,
vars: HashMap<Variable, Set>,
choice_instance: &ChoiceInstance
) -> (Vec<Set>, SetConstraints, Adaptator)

Generates the list of sets to iterate and to constraints to iterate on the given context, but from the point of view of the given choice instance.

pub fn adapt_env_ext(
&self,
vars: HashMap<Variable, Set>,
choice_instance: &ChoiceInstance
) -> (Vec<Set>, Vec<Set>, SetConstraints, Adaptator)

Generates the foralls and the set constraints to iterate on the given environment, from the point of view of the given choice. Returns the foralls issued from arguments in a different list than the foralls issued from foralls in the original environment.