Struct telamon_gen::ir::Set

pub struct Set { /* fields omitted */ }

References a set of objects.

Methods

impl Set

pub fn new(def: &Rc<SetDef>, var: Option<Variable>) -> Self

Create a new set instance.

pub fn is_subset_of_def(&self, other: &Set) -> bool

Indicates if the first set is a sub-set of the second, without matching argument names.

pub fn get_collision_level<'a>(&'a self, rhs: &Set) -> Option<&'a SetDef>

Returns the common superset where the two set might have an object in common.

pub fn reverse(&self, self_var: Variable, arg: &Set) -> Option<(Set, Set)>

Returns a superset of this set and a set parametrized by elements of the superset that iterates on the possible parameters of this set given a variable of the superset.