Module telamon_gen::lexer

Structs

Lexer

LexerPosition
A sequence's row/column position

LexicalError

Position

Span
A double sequence's row/column position

Spanned

Enums

ErrorKind

Token

Type Definitions

LexerItem

ParserSpanned
The alias Spanned is a definition of the stream format. The parser will accept an iterator where each item in the stream has the following structure.