Enum telamon_gen::ir::Condition

pub enum Condition {
    Code {
        code: Code,
        negate: bool,
    },
    Enum {
        input: usize,
        values: BTreeSet<RcStr>,
        negate: bool,
        inverse: bool,
    },
    Bool(bool),
    CmpCode {
        lhs: usize,
        rhs: Code,
        op: CmpOp,
    },
    CmpInput {
        lhs: usize,
        rhs: usize,
        op: CmpOp,
        inverse: bool,
    },
}

A condition producing a boolean.

Variants
Code

Triggers if the given piece of code returns true, given the mapping of variables.

Fields of Code

code: Code

negate: bool

Enum

Triggers if the choice can only take the given values.

Fields of Enum

input: usize

values: BTreeSet<RcStr>

negate: bool

inverse: bool

Bool(bool)

Always or never triggers.

CmpCode

Triggers if the one inputs respects the comparison operator with some rust code.

Fields of CmpCode

lhs: usize

rhs: Code

op: CmpOp

CmpInput

Triggers if the two inputs respects the comparison operator.

Fields of CmpInput

lhs: usize

rhs: usize

op: CmpOp

inverse: bool

Methods

impl Condition

pub fn negate(&mut self)

Negates the condition.

pub fn alternatives_of(
&self,
input_id: usize,
t: &ValueType,
ir_desc: &IrDesc
) -> Option<ValueSet>

Returns allowed alternatives for the given input. Returns None if the condition is not on the given input.

pub fn instantiate(
&self,
inputs: &[ChoiceInstance],
input_mapping: &HashMap<usize, &ValueSet>,
ir_desc: &IrDesc
) -> Condition

Instantiate the condition in the given context.

pub fn evaluate(
&self,
inputs: &[ChoiceInstance],
input_mapping: &HashMap<usize, &ValueSet>,
ir_desc: &IrDesc
) -> Trivalent

Evaluates the condition. Requires the mapping to be instantiated.

pub fn normalize(&mut self, inputs: &[ChoiceInstance], ir_desc: &IrDesc)

Normalizes the condition to make it easier to apply equality on it.