Enum telamon_gen::ir::ValueSet

pub enum ValueSet {
    Enum {
        enum_name: RcStr,
        values: BTreeSet<RcStr>,
        inputs: BTreeSet<(usize, bool, bool)>,
    },
    Integer {
        is_full: bool,
        cmp_inputs: BTreeSet<(CmpOp, usize)>,
        cmp_code: BTreeSet<(CmpOp, Code)>,
        universe: ValueType,
    },
}

Represents a set of values a choice can take.

Variants
Enum
Fields of Enum

enum_name: RcStr

The enum type.

values: BTreeSet<RcStr>

A fixed set of values. Does not contains aliases.

inputs: BTreeSet<(usize, bool, bool)>

A set of inputs whose values can be take by the choice. Two flags indicate if the value should be negated or inversed.

Integer
Fields of Integer

is_full: bool

cmp_inputs: BTreeSet<(CmpOp, usize)>

cmp_code: BTreeSet<(CmpOp, Code)>

universe: ValueType

Methods

impl ValueSet

pub fn empty(t: &ValueType) -> Self

Creates an enmpty ValueSet of the given type.

pub fn from_properties(
t: &ValueType,
is_eq: bool,
is_inv: bool,
ir_desc: &IrDesc
) -> Self

Computes a ValueSet from the properties it must respect.

pub fn enum_values(enum_: RcStr, values: BTreeSet<RcStr>) -> Self

Creates a ValueSet from a normalized set of values.

pub fn from_input(t: &ValueType, input: usize, op: CmpOp, inverse: bool) -> Self

Creates a ValueSet from the given input.

pub fn is_empty(&self) -> bool

Indicates if the set of values is empty.

pub fn is_full(&self, ir_desc: &IrDesc) -> bool

Indicates if the set contains all the values. This functions is pessimistic: the set may contain all the values and the function still return false.

pub fn is_constrained(&self) -> Trivalent

Indicates if the set contains a single value.

pub fn extend(&mut self, other: ValueSet)

Extends the ValueSet with the values of anther set.

pub fn intersect(&mut self, other: ValueSet) -> bool

Intersects the ValueSet with the given values. Indicates if the intersection was successful or if the sets should be kept separate.

pub fn instantiate(
&self,
input_mapping: &HashMap<usize, &ValueSet>,
ir_desc: &IrDesc
) -> Self

Instantiates the ValueSet for a given input assignment.

pub fn inverse(&mut self, ir_desc: &IrDesc)

Inverse the ValueSet. The choice must be antisymmetric.

pub fn is(&self, other: &ValueSet) -> Trivalent

Indicates if the ValueSet will be contained into anoter after instantiation. Requires both self and other to be instantiated.

pub fn t(&self) -> ValueType

Returns the type of the values.