Struct telamon_gen::ir::Choice

pub struct Choice { /* fields omitted */ }

A decision to specify.

Methods

impl Choice

pub fn new(
name: RcStr,
doc: Option<RcStr>,
arguments: ChoiceArguments,
def: ChoiceDef
) -> Self

Creates a new Choice.

pub fn name(&self) -> &RcStr

Returns the name of the choice, in snake_case.

pub fn doc(&self) -> Option<&str>

Returns the documentation associated with the Choice.

pub fn arguments(&self) -> &ChoiceArguments

Returns the parameters for which the Choice is defined.

pub fn value_type(&self) -> ValueType

Returns the type representing the values the Choice can take.

pub fn choice_def(&self) -> &ChoiceDef

Returns the definition of the `Choice.

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> type Item = &'a T;

pub fn on_change(&self) -> Iter<OnChangeAction>

Returns the actions to perform when the Choice is constrained.

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> type Item = &'a T;

pub fn filter_actions(&self) -> Iter<FilterAction>

Returns the actions to run to get the valid alternatives of the choice.

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> type Item = &'a T;

pub fn filters(&self) -> Iter<Filter>

Returns the filters operating on the Choice.

pub fn add_filter_action(&mut self, action: FilterAction)

Adds a filter to run on initialization.

pub fn add_onchange(&mut self, action: OnChangeAction)

Adds an action to perform when the Choice is constrained.

pub fn add_filter(&mut self, filter: Filter) -> usize

Adds a filter to the Choice, returns an ID to indentify it.

pub fn fragile_values(&self) -> &ValueSet

Returns the values that should not be automatically restricted by filters.

pub fn add_fragile_values(&mut self, values: ValueSet)

Extends the list of values that should not be automatically propagated by filters.