Trait telamon_gen::ir::SetRef

pub trait SetRef<'a> {
    fn def(&self) -> &'a SetDef;

    fn arg(&self) -> Option<Variable>;

    fn reverse_constraint(&self) -> Option<SetRefImpl<'a>>;

    fn without_reverse_constraints(&self) -> SetRefImpl<'a> { ... }

    fn superset(&self) -> Option<SetRefImpl<'a>> { ... }

    fn path_to_superset(&self, superset: &SetRef) -> Vec<SetRefImpl<'a>> { ... }

    fn is_subset_of(&self, other: &Set) -> bool { ... }

    fn as_ref(&self) -> SetRefImpl<'a> { ... }
}

Generic trait for sets.

Required Methods

fn def(&self) -> &'a SetDef

Returns the set definition.

fn arg(&self) -> Option<Variable>

Returns the argument of the set, if any.

fn reverse_constraint(&self) -> Option<SetRefImpl<'a>>

A constraint on the variables to iterate on, issued from a set reversal.

Provided Methods

fn without_reverse_constraints(&self) -> SetRefImpl<'a>

Returns the same set but without reverse constraints.

fn superset(&self) -> Option<SetRefImpl<'a>>

Returns the direct superset of this set, if any.

Important traits for Vec<u8>
impl Write for Vec<u8>

fn path_to_superset(&self, superset: &SetRef) -> Vec<SetRefImpl<'a>>

Returns the path of sets to access a super-set.

fn is_subset_of(&self, other: &Set) -> bool

Indicates if the first set is a sub-set of the second.

fn as_ref(&self) -> SetRefImpl<'a>

Returns the SetRefImpl corresponding to this set.